+++ src/libsac2c/global/ctinfo.h
@@ -21,7 +21,7 @@ extern void CTIabortOnBottom (char *err_msg);
 extern void CTIabort (const char *format, ...) PRINTF_FORMAT (1, 2) FUN_ATTR_NORETURN;
 extern void CTIabortLine (size_t line, const char *format, ...)
   PRINTF_FORMAT (2, 3) FUN_ATTR_NORETURN;
-extern void CTIabortOutOfMemory (unsigned int request) FUN_ATTR_NORETURN;
+extern void CTIabortOutOfMemory (size_t request) FUN_ATTR_NORETURN;
 extern void CTIabortOnError (void);
 extern void CTIwarn (const char *format, ...) PRINTF_FORMAT (1, 2);
 extern void CTIwarnLoc (struct location loc, const char *format, ...)
